feat: 헤딩 해제 기능 추가 및 에디터 유틸/훅 분리
