+++ src/components/inquiry/form/InquiryEditor.tsx
@@ -1,6 +1,8 @@
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useState } from "react";
 
 import { toolbarButtons } from "@/constants/toolbarButtons";
+import { useEditor } from "@/hooks/useEditor";
+import { isToolbarItem } from "@/utils/commandUtils";
 import { EditorCommand } from "@/types/toolbar";
 
 import "@/styles/editor.css";
@@ -8,92 +10,14 @@ import "@toast-ui/editor/dist/toastui-editor.css";
 import { Editor } from "@toast-ui/react-editor";
 
 const InquiryEditor = () => {
-  const editorRef = useRef<Editor>(null);
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
-  const [activeSet, setActiveSet] = useState<Set<string>>(new Set());
+  const { editorRef, fileInputRef, activeSet, execCommand, handleFileChange } =
+    useEditor();
   const [title, setTitle] = useState<string>("");
-
-  const isToolbarItem = (
-    item: (typeof toolbarButtons)[number]
-  ): item is Exclude<typeof item, "divider"> => {
-    return typeof item !== "string";
-  };
-
-  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-
-    const imageUrl = URL.createObjectURL(file);
-    const editorInstance = editorRef.current?.getInstance();
-    editorInstance?.insertText(`![이미지 설명](${imageUrl})`);
-
-    e.target.value = "";
-  };
-
-  const isInlineCommand = (
-    cmd: EditorCommand
-  ): cmd is {
-    command: "bold" | "italic" | "strike" | "link" | "image";
-  } => {
-    return cmd.command !== "heading" && cmd.command !== "quote";
-  };
-
-  const isHeadingCommand = (
-    cmd: EditorCommand
-  ): cmd is { command: "heading"; payload: { level: number } } => {
-    return cmd.command === "heading";
-  };
-
-  const execCommand = (params: EditorCommand) => {
-    const editorInstance = editorRef.current?.getInstance();
-    if (!editorInstance) return;
-
-    const newSet = new Set(activeSet);
-
-    if (isInlineCommand(params)) {
-      const isActive = newSet.has(params.command);
-
-      editorInstance.exec(params.command);
-
-      if (isActive) {
-        newSet.delete(params.command);
-      } else {
-        newSet.add(params.command);
-      }
-    } else if (params.command === "quote") {
-      const isActive = newSet.has("quote");
-
-      if (isActive) {
-        editorInstance.exec("removeBlockQuote");
-        newSet.delete("quote");
-      } else {
-        editorInstance.exec("blockQuote");
-        newSet.add("quote");
-      }
-    } else if (isHeadingCommand(params)) {
-      const current = `h${params.payload.level}`;
-      const isActive = newSet.has(current);
-
-      if (isActive) {
-        editorInstance.exec("paragraph");
-        ["h1", "h2", "h3", "h4"].forEach(h => newSet.delete(h));
-      } else {
-        editorInstance.exec("heading", params.payload);
-        ["h1", "h2", "h3", "h4"].forEach(h => newSet.delete(h));
-        newSet.add(current);
-      }
-    }
-
-    setActiveSet(newSet);
-  };
 
   useEffect(() => {
     const editorInstance = editorRef.current?.getInstance();
-    if (editorInstance) {
-      editorInstance.setMarkdown("");
-    }
-  }, []);
+    editorInstance?.setMarkdown("");
+  }, [editorRef]);
 
   return (
     <div className="flex flex-col gap-6">
@@ -131,7 +55,12 @@ const InquiryEditor = () => {
                     payload: { level: item.level },
                   });
                 } else {
-                  execCommand({ command: item.command } as EditorCommand);
+                  execCommand({
+                    command: item.command as Exclude<
+                      EditorCommand["command"],
+                      "heading"
+                    >,
+                  });
                 }
               }}
               className={`cursor-pointer ${
